feat: update subscriptions example
