+++ package.json
@@ -41,6 +41,7 @@
     "examples/fastify",
     "examples/graphql-config",
     "examples/nextjs",
+    "examples/subscriptions",
     "website"
   ],
   "homepage": "https://github.com/dotansimha/graphql-yoga#readme",
